rename: 프로젝트 패키지명 변경
- 이전의 실수로 인해 패키지명이 src 로 설정되어 있었음.
- lotto 로 변경함으로써 로또 프로젝트라는 것을 명시.
