+++ src/main/java/lotto/Application.java
@@ -1,0 +1,12 @@
+package lotto;
+
+import lotto.controller.LottoController;
+
+public class Application {
+
+    private static final LottoController lottoController = new LottoController();
+
+    public static void main(String[] args) {
+        lottoController.run();
+    }
+}
