+++ src/main/java/lotto/model/lotto/generator/NumberGenerator.java
@@ -1,0 +1,8 @@
+package lotto.model.lotto.generator;
+
+import java.util.List;
+
+public interface NumberGenerator {
+
+    List<Integer> generate(int min, int max, int size);
+}
