+++ src/main/java/lotto/view/input/InputView.java
@@ -1,0 +1,12 @@
+package lotto.view.input;
+
+import java.util.List;
+
+public interface InputView {
+
+    int inputPurchaseMoney();
+
+    List<Integer> inputWinningLottoNumbers();
+
+    int inputBonusNumber();
+}
