+++ src/main/java/lotto/view/output/OutputView.java
@@ -1,0 +1,16 @@
+package lotto.view.output;
+
+import java.util.List;
+
+public interface OutputView {
+
+    void printInputPurchaseMoneyMessage();
+
+    void printPurchasedLottos(List<LottoResponse> responses);
+
+    void printInputWinningLottoNumbers();
+
+    void printInputBonusNumber();
+
+    void printLottoResults(List<LottoPrizeResponse> lottoPrizeResponses, int purchaseMoney);
+}
