--- src/main/java/src/Application.java
@@ -1,12 +1,0 @@
-package src;
-
-import src.controller.LottoController;
-
-public class Application {
-
-    private static final LottoController lottoController = new LottoController();
-
-    public static void main(String[] args) {
-        lottoController.run();
-    }
-}
