--- src/main/java/src/model/lotto/generator/NumberGenerator.java
@@ -1,8 +1,0 @@
-package src.model.lotto.generator;
-
-import java.util.List;
-
-public interface NumberGenerator {
-
-    List<Integer> generate(int min, int max, int size);
-}
