--- src/main/java/src/view/input/InputView.java
@@ -1,12 +1,0 @@
-package src.view.input;
-
-import java.util.List;
-
-public interface InputView {
-
-    int inputPurchaseMoney();
-
-    List<Integer> inputWinningLottoNumbers();
-
-    int inputBonusNumber();
-}
