--- src/main/java/src/view/output/OutputView.java
@@ -1,16 +1,0 @@
-package src.view.output;
-
-import java.util.List;
-
-public interface OutputView {
-
-    void printInputPurchaseMoneyMessage();
-
-    void printPurchasedLottos(List<LottoResponse> responses);
-
-    void printInputWinningLottoNumbers();
-
-    void printInputBonusNumber();
-
-    void printLottoResults(List<LottoPrizeResponse> lottoPrizeResponses, int purchaseMoney);
-}
