+++ src/test/java/lotto/ApplicationTest.java
@@ -1,0 +1,4 @@
+package lotto;
+
+public class ApplicationTest {
+}
