--- src/test/java/src/ApplicationTest.java
@@ -1,4 +1,0 @@
-package src;
-
-public class ApplicationTest {
-}
