--- src/test/java/src/model/winning_lotto/vo/WinningLottoNumberTest.java
@@ -1,18 +1,0 @@
-package src.model.winning_lotto.vo;
-
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static src.model.winning_lotto.vo.WinningLottoNumber.MAX;
-import static src.model.winning_lotto.vo.WinningLottoNumber.MIN;
-
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
-
-class WinningLottoNumberTest {
-
-    @ParameterizedTest
-    @ValueSource(ints = {MIN - 1, MAX + 1})
-    void 당첨_번호_생성_시_최대_최소_범위를_넘어가면_예외_발생(int value) {
-
-        assertThatIllegalArgumentException().isThrownBy(() -> WinningLottoNumber.from(value));
-    }
-}
